perbaiki fitur regis & url

+++ admin/hapus_kategori.php
@@ -1,5 +1,12 @@
 <?php 
+require '../config/functions.php';
 require 'layout/header.php';
+
+// mengatasi iseng menulis url
+if(!isset($_GET['id'])) {
+	header("Location: " . base_url('admin/list_kategori.php'));
+	exit;
+}
 
 $id = $_GET['id'];
 mysqli_query($conn, "DELETE FROM tb_kategori WHERE id_kategori = $id");
